Se agregaron todas las tablas de los datos extras de personas

+++ src/models/persona_models/persona_beneficiario_models.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, sqlx::FromRow, Deserialize, Serialize)]
 pub struct BeneficiarioPersonaModelo {
     pub id_persona_beneficiario: i32,
-    pub id_persona: i32,
-    pub entre_calle: i32,
+    pub id_persona: i64,
+    pub entre_calle: i64,
     pub y_calle: i32,
 }
